+++ client_web_app/src/App.css
@@ -5,4 +5,25 @@
   align-items: center;
   flex-direction: column;
 }
+.fileTable {
+  font-family: Arial, Helvetica, sans-serif;
+  border-collapse: collapse;
+  width: 100%;
+}
 
+.fileTable td, .fileTable th {
+  border: 1px solid #ddd;
+  padding: 8px;
+}
+
+.fileTable tr:nth-child(even){background-color: #f2f2f2;}
+
+.fileTable tr:hover {background-color: #ddd;}
+
+.fileTable th {
+  padding-top: 12px;
+  padding-bottom: 12px;
+  text-align: left;
+  background-color: #04AA6D;
+  color: white;
+}
